[#690] - Mueve parsing de contenido de Story y StoryPreview a backend
