+++ src/api/_utils/functions.ts
@@ -13,6 +13,7 @@ import { baseLanguage } from '../../../cms/utils/localization';
 import { AuthorDTO } from '@models/author.model';
 import { mapMediaSources } from './media-sources.functions';
 import { StorylistDTO } from '@models/storylist.model';
+import { Story, StoryPreview } from '@models/story.model';
 
 export function mapAuthor(rawAuthorData: any, language?: string): AuthorDTO {
 	return {
@@ -75,13 +76,13 @@ export async function mapStorylist(result: any): Promise<StorylistDTO> {
 		const { review, body, author, mediaSources, ...story } = publication.story;
 		publications.push({
 			...publication,
-			story: {
+			story: mapStoryPreviewContent({
 				...story,
 				media: mediaSources ? await mapMediaSources(mediaSources) : undefined,
 				summary: review,
 				paragraphs: body,
 				author: mapAuthorForStory(author),
-			},
+			}),
 		});
 	}
 
@@ -97,4 +98,34 @@ export async function mapStorylist(result: any): Promise<StorylistDTO> {
 					})),
 		publications: publications,
 	};
+}
+
+export function mapStoryContent(story: Story): Story {
+	return {
+		...story,
+		epigraphs: story.epigraphs ?? [],
+		paragraphs: story?.paragraphs ?? [],
+		summary: story?.summary ?? [],
+		author: {
+			...story.author,
+			biography: story.author.biography ?? [],
+		},
+		media: story.media ?? [],
+	};
+}
+
+export function mapStoryPreviewContent(story: StoryPreview): StoryPreview {
+	const card = {
+		...story,
+		paragraphs: story?.paragraphs ?? [],
+		media: story.media ?? [],
+		originalPublication: story.originalPublication ?? '',
+	};
+
+	if (story.author) {
+		card.author = {
+			...story.author,
+		};
+	}
+	return card;
 }
